requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
Django==4.2
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
django-tinymce==3.6.1
Pillow==9.5.0
sqlparse==0.4.3
tzdata==2023.3
